Moving legacy dylib/vmvx driver modules to iree/hal/drivers/. These will be replaced with local-* variants in the future. Progress on #9336.
