+++ runtime/src/iree/hal/drivers/CMakeLists.txt
@@ -12,16 +12,20 @@ if(IREE_HAL_DRIVER_CUDA)
   list(APPEND IREE_HAL_DRIVER_MODULES iree::hal::drivers::cuda::registration)
 endif()
 if(IREE_HAL_DRIVER_DYLIB)
-  list(APPEND IREE_HAL_DRIVER_MODULES iree::hal::dylib::registration)
+  add_subdirectory(dylib)
+  list(APPEND IREE_HAL_DRIVER_MODULES iree::hal::drivers::dylib::registration)
 endif()
 if(IREE_HAL_DRIVER_DYLIB_SYNC)
-  list(APPEND IREE_HAL_DRIVER_MODULES iree::hal::dylib::registration::sync)
+  add_subdirectory(dylib_sync)
+  list(APPEND IREE_HAL_DRIVER_MODULES iree::hal::drivers::dylib_sync::registration)
 endif()
 if(IREE_HAL_DRIVER_VMVX)
-  list(APPEND IREE_HAL_DRIVER_MODULES iree::hal::vmvx::registration)
+  add_subdirectory(vmvx)
+  list(APPEND IREE_HAL_DRIVER_MODULES iree::hal::drivers::vmvx::registration)
 endif()
 if(IREE_HAL_DRIVER_VMVX_SYNC)
-  list(APPEND IREE_HAL_DRIVER_MODULES iree::hal::vmvx::registration::sync)
+  add_subdirectory(vmvx_sync)
+  list(APPEND IREE_HAL_DRIVER_MODULES iree::hal::drivers::vmvx_sync::registration)
 endif()
 if(IREE_HAL_DRIVER_VULKAN)
   add_subdirectory(vulkan)
