+++ runtime/src/iree/hal/drivers/dylib/registration/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# runtime/src/iree/hal/dylib/registration/BUILD                                #
+# runtime/src/iree/hal/drivers/dylib/registration/BUILD                        #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -9,8 +9,6 @@
 ################################################################################
 
 iree_add_all_subdirs()
-
-if(IREE_HAL_DRIVER_DYLIB)
 
 iree_cc_library(
   NAME
@@ -32,29 +30,5 @@ iree_cc_library(
     "IREE_HAL_HAVE_DYLIB_DRIVER_MODULE=1"
   PUBLIC
 )
-
-endif()
-
-if(IREE_HAL_DRIVER_DYLIB_SYNC)
-
-iree_cc_library(
-  NAME
-    sync
-  HDRS
-    "driver_module_sync.h"
-  SRCS
-    "driver_module_sync.c"
-  DEPS
-    iree::base
-    iree::hal
-    iree::hal::local
-    iree::hal::local::loaders::embedded_library_loader
-    iree::hal::local::sync_driver
-  DEFINES
-    "IREE_HAL_HAVE_DYLIB_SYNC_DRIVER_MODULE=1"
-  PUBLIC
-)
-
-endif()
 
 ### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
