+++ runtime/src/iree/hal/drivers/dylib_sync/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# runtime/src/iree/hal/vmvx/BUILD                                              #
+# runtime/src/iree/hal/drivers/dylib_sync/BUILD                                #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
