+++ runtime/src/iree/hal/drivers/dylib_sync/cts/CMakeLists.txt
@@ -14,24 +14,9 @@ set(EXECUTABLE_FORMAT "\"${EXECUTABLE_FORMAT_PREFIX}-elf-\" IREE_ARCH")
 
 iree_hal_cts_test_suite(
   DRIVER_NAME
-    dylib
-  DRIVER_REGISTRATION_HDR
-    "runtime/src/iree/hal/dylib/registration/driver_module.h"
-  DRIVER_REGISTRATION_FN
-    "iree_hal_dylib_driver_module_register"
-  COMPILER_TARGET_BACKEND
-    "dylib-llvm-aot"
-  EXECUTABLE_FORMAT
-    "${EXECUTABLE_FORMAT}"
-  DEPS
-    iree::hal::dylib::registration
-)
-
-iree_hal_cts_test_suite(
-  DRIVER_NAME
     dylib-sync
   DRIVER_REGISTRATION_HDR
-    "runtime/src/iree/hal/dylib/registration/driver_module_sync.h"
+    "runtime/src/iree/hal/drivers/dylib_sync/registration/driver_module.h"
   DRIVER_REGISTRATION_FN
     "iree_hal_dylib_sync_driver_module_register"
   COMPILER_TARGET_BACKEND
@@ -39,7 +24,7 @@ iree_hal_cts_test_suite(
   EXECUTABLE_FORMAT
     "${EXECUTABLE_FORMAT}"
   DEPS
-    iree::hal::dylib::registration::sync
+    iree::hal::drivers::dylib_sync::registration
   EXCLUDED_TESTS
     # TODO(#4680): command buffer recording so that these can run on sync HAL
     "command_buffer"
