+++ runtime/src/iree/hal/drivers/dylib_sync/registration/driver_module.c
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/hal/dylib/registration/driver_module_sync.h"
+#include "iree/hal/drivers/dylib_sync/registration/driver_module.h"
 
 #include <inttypes.h>
 #include <stddef.h>
