+++ runtime/src/iree/hal/drivers/dylib_sync/registration/driver_module.h
@@ -4,8 +4,8 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef IREE_HAL_DYLIB_REGISTRATION_DRIVER_MODULE_SYNC_H_
-#define IREE_HAL_DYLIB_REGISTRATION_DRIVER_MODULE_SYNC_H_
+#ifndef IREE_HAL_DRIVER_DYLIB_SYNC_REGISTRATION_DRIVER_MODULE_H_
+#define IREE_HAL_DRIVER_DYLIB_SYNC_REGISTRATION_DRIVER_MODULE_H_
 
 #include "iree/base/api.h"
 #include "iree/hal/api.h"
@@ -23,4 +23,4 @@ IREE_API_EXPORT iree_status_t iree_hal_dylib_sync_driver_module_register(
 }  // extern "C"
 #endif  // __cplusplus
 
-#endif  // IREE_HAL_DYLIB_REGISTRATION_DRIVER_MODULE_SYNC_H_
+#endif  // IREE_HAL_DRIVER_DYLIB_SYNC_REGISTRATION_DRIVER_MODULE_H_
