+++ runtime/src/iree/hal/drivers/vmvx_sync/cts/CMakeLists.txt
@@ -6,24 +6,9 @@
 
 iree_hal_cts_test_suite(
   DRIVER_NAME
-    vmvx
-  DRIVER_REGISTRATION_HDR
-    "runtime/src/iree/hal/vmvx/registration/driver_module.h"
-  DRIVER_REGISTRATION_FN
-    "iree_hal_vmvx_driver_module_register"
-  COMPILER_TARGET_BACKEND
-    "vmvx"
-  EXECUTABLE_FORMAT
-    "\"vmvx-bytecode-fb\""
-  DEPS
-    iree::hal::vmvx::registration
-)
-
-iree_hal_cts_test_suite(
-  DRIVER_NAME
     vmvx-sync
   DRIVER_REGISTRATION_HDR
-    "runtime/src/iree/hal/vmvx/registration/driver_module_sync.h"
+    "runtime/src/iree/hal/drivers/vmvx_sync/registration/driver_module.h"
   DRIVER_REGISTRATION_FN
     "iree_hal_vmvx_sync_driver_module_register"
   COMPILER_TARGET_BACKEND
@@ -31,7 +16,7 @@ iree_hal_cts_test_suite(
   EXECUTABLE_FORMAT
     "\"vmvx-bytecode-fb\""
   DEPS
-    iree::hal::vmvx::registration::sync
+    iree::hal::drivers::vmvx_sync::registration
   EXCLUDED_TESTS
     # TODO(#4680): command buffer recording so that these can run on sync HAL
     "command_buffer"
