+++ runtime/src/iree/hal/drivers/vmvx_sync/registration/driver_module.h
@@ -4,8 +4,8 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef IREE_HAL_VMVX_REGISTRATION_DRIVER_MODULE_H_
-#define IREE_HAL_VMVX_REGISTRATION_DRIVER_MODULE_H_
+#ifndef IREE_HAL_DRIVER_VMVX_SYNC_REGISTRATION_DRIVER_MODULE_H_
+#define IREE_HAL_DRIVER_VMVX_SYNC_REGISTRATION_DRIVER_MODULE_H_
 
 #include "iree/base/api.h"
 #include "iree/hal/api.h"
@@ -14,11 +14,13 @@
 extern "C" {
 #endif  // __cplusplus
 
+// DEPRECATED: this entire driver will be removed soon.
+// TODO(#3580): remove this entire driver w/ iree_hal_executable_library_t.
 IREE_API_EXPORT iree_status_t
-iree_hal_vmvx_driver_module_register(iree_hal_driver_registry_t* registry);
+iree_hal_vmvx_sync_driver_module_register(iree_hal_driver_registry_t* registry);
 
 #ifdef __cplusplus
 }  // extern "C"
 #endif  // __cplusplus
 
-#endif  // IREE_HAL_VMVX_REGISTRATION_DRIVER_MODULE_H_
+#endif  // IREE_HAL_DRIVER_VMVX_SYNC_REGISTRATION_DRIVER_MODULE_H_
